add a --role-arn argument
This will give the user the ability to pass in the ARN of the role
he/she is attempting to assume the credentials for.  This is useful if
the user has access to multiple roles he/she can assume, and saves them
from being prompted to select one from a list.